Deprecated header-link and clone-attribute scripts.
header-link.js and clone-attribute.js have been replaced with a Metalsmith plugin that runs at build time, reducing the impact on the client's machine. Also, metalsmith-beautify was dropped in favour of metalsmith-html-minimizer and metalsmith-clean-css to further reduce load time at the cost of raw source readability in the generated files.